requirements.txt
python-dotenv>=1.0.0
pytest>=8.0
pytest-timeout>=2.2
